Extract ParseClusterConfig fn to /pkg
